requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
async-timeout==4.0.3
openai>=1.40.0
python-dotenv==1.0.1
aiocryptopay==0.4.0
requests>=2.31.0
httpx>=0.27.0